+++ Package.swift
@@ -9,6 +9,7 @@ import PackageDescription
 
 let package = Package(
     name: "FeedbackThread",
+    defaultLocalization: "en",
     platforms: [
         .iOS(.v16),
         .macOS(.v13),